Add more parameters to eeprom

+++ inc/parameters.h
@@ -5,18 +5,18 @@
 
 typedef struct sumo_parameters {
 
-    uint8_t current_strategy;
-    uint8_t current_pre_strategy;
+    uint8_t  current_strategy;
+    uint8_t  current_pre_strategy;
 
-    uint8_t enabled_distance_sensors;
-    uint8_t enabled_line_sensors;
+    uint16_t enabled_distance_sensors;
+    uint8_t  enabled_line_sensors;
 
-    uint8_t star_speed;
-    uint8_t max_speed;
+    uint8_t  star_speed;
+    uint8_t  max_speed;
     
-    uint8_t reverse_speed; 
+    uint8_t  reverse_speed; 
     uint16_t reverse_time_ms;
-    uint8_t turn_speed;
+    uint8_t  turn_speed;
 
     uint16_t turn_180_time_ms;
